+++ Changelog.md
@@ -1,5 +1,11 @@
 # Vanilo Cloud PHP SDK Changelog
 
+## Unreleased
+##### 2025-XX-YY
+
+- Added support for transposing arrays of objects
+- Added `OrderItem` as the type of the elements of the `items` array in the `Order` model
+
 ## 0.8.0
 ##### 2025-02-21